refactor(ods): standardize deploy-target resolution across deploy subcommands
Make `ods deploy wiki` resolve its deploy target (the repo hosting the
deploy workflow + the workflow filename) the same way `ods deploy edge`
does: explicit flags, then saved config, then a first-run prompt persisted
to the ods config file. Previously wiki hardcoded its target while edge
made it configurable.

- Generalize the persisted config type into a shared DeployTarget and add a
  deploy_wiki section (on-disk JSON tags unchanged, so existing configs work)
- Extract the shared resolver and gh workflow plumbing into deploy_common.go
- Add --target-repo / --target-workflow flags to `ods deploy wiki`
