+++ tools/ods/cmd/deploy_wiki.go
@@ -6,6 +6,7 @@ import (
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
+	"github.com/onyx-dot-app/onyx/tools/ods/internal/config"
 	"github.com/onyx-dot-app/onyx/tools/ods/internal/git"
 	"github.com/onyx-dot-app/onyx/tools/ods/internal/prompt"
 )
@@ -13,18 +14,18 @@ import (
 const (
 	wikiBuildRepo       = "onyx-dot-app/agent-wiki"
 	wikiBuildWorkflow   = "nightly-build.yml"
-	wikiDeployRepo      = "onyx-dot-app/cloud-deployment-yamls"
-	wikiDeployWorkflow  = "dev-wiki-deploy.yml"
 	wikiBuildPollLimit  = 30 * time.Minute
 	wikiDeployPollLimit = 20 * time.Minute
 )
 
 // DeployWikiOptions holds options for the deploy wiki command.
 type DeployWikiOptions struct {
-	DryRun       bool
-	Yes          bool
-	NoWaitDeploy bool
-	NoBuild      bool
+	TargetRepo     string
+	TargetWorkflow string
+	DryRun         bool
+	Yes            bool
+	NoWaitDeploy   bool
+	NoBuild        bool
 }
 
 // NewDeployWikiCommand creates the `ods deploy wiki` command.
@@ -40,13 +41,18 @@ This command will:
   1. Dispatch the nightly-build.yml workflow in onyx-dot-app/agent-wiki
      (builds and pushes onyxdotapp/agent-wiki-{backend,frontend}:nightly-latest-YYYYMMDD)
   2. Wait for the build workflow to finish
-  3. Dispatch the dev-wiki-deploy.yml workflow in onyx-dot-app/cloud-deployment-yamls
-     with version_tag=nightly-latest-YYYYMMDD (today's UTC date)
+  3. Dispatch the configured deploy workflow with version_tag=nightly-latest-YYYYMMDD
+     (today's UTC date)
   4. Wait for the deploy workflow to finish
 
 All GitHub operations run through the gh CLI, so authorization is enforced
 by your gh credentials and GitHub's repo/workflow permissions. A kickoff
 Slack message will appear in #monitor-deployments.
+
+On first run, you'll be prompted for the deploy target repo and workflow
+filename. These are saved to the ods config file (~/.config/onyx-dev/config.json
+on Linux/macOS) and reused on subsequent runs. Pass --target-repo or
+--target-workflow to override the saved values.
 
 Pass --no-build to skip step 1 and just deploy whatever's already on
 Docker Hub for today's tag.
@@ -60,6 +66,8 @@ Example usage:
 		},
 	}
 
+	cmd.Flags().StringVar(&opts.TargetRepo, "target-repo", "", "GitHub repo (owner/name) hosting the deploy workflow; overrides saved config")
+	cmd.Flags().StringVar(&opts.TargetWorkflow, "target-workflow", "", "Filename of the deploy workflow within the target repo; overrides saved config")
 	cmd.Flags().BoolVar(&opts.DryRun, "dry-run", false, "Perform local operations only; skip dispatching workflows")
 	cmd.Flags().BoolVar(&opts.Yes, "yes", false, "Skip the confirmation prompt")
 	cmd.Flags().BoolVar(&opts.NoWaitDeploy, "no-wait-deploy", false, "Do not wait for the deploy workflow to finish after dispatching it")
@@ -70,6 +78,12 @@ Example usage:
 
 func deployWiki(opts *DeployWikiOptions) {
 	git.CheckGitHubCLI()
+
+	deployRepo, deployWorkflow := resolveDeployTarget(
+		opts.TargetRepo,
+		opts.TargetWorkflow,
+		func(c *config.Config) *config.DeployTarget { return &c.DeployWiki },
+	)
 
 	if opts.DryRun {
 		log.Warning("=== DRY RUN MODE: workflow dispatches will be skipped ===")
@@ -100,11 +114,11 @@ func deployWiki(opts *DeployWikiOptions) {
 	}
 
 	if opts.DryRun {
-		log.Warnf("[DRY RUN] Would dispatch %s in %s with version_tag=%s", wikiDeployWorkflow, wikiDeployRepo, versionTag)
+		log.Warnf("[DRY RUN] Would dispatch %s in %s with version_tag=%s", deployWorkflow, deployRepo, versionTag)
 		return
 	}
 
-	runDeploy(versionTag, opts.NoWaitDeploy)
+	runDeploy(deployRepo, deployWorkflow, versionTag, opts.NoWaitDeploy)
 }
 
 func runBuild() {
@@ -132,20 +146,20 @@ func runBuild() {
 	log.Info("Build completed successfully.")
 }
 
-func runDeploy(versionTag string, noWait bool) {
-	priorRunID, err := latestWorkflowRunID(wikiDeployRepo, wikiDeployWorkflow, "workflow_dispatch", "")
+func runDeploy(deployRepo, deployWorkflow, versionTag string, noWait bool) {
+	priorRunID, err := latestWorkflowRunID(deployRepo, deployWorkflow, "workflow_dispatch", "")
 	if err != nil {
 		log.Fatalf("Failed to query existing deploy runs: %v", err)
 	}
 	log.Debugf("Most recent prior deploy run id: %d", priorRunID)
 
-	log.Infof("Dispatching %s with version_tag=%s...", wikiDeployWorkflow, versionTag)
-	if err := dispatchWorkflow(wikiDeployRepo, wikiDeployWorkflow, map[string]string{"version_tag": versionTag}); err != nil {
+	log.Infof("Dispatching %s with version_tag=%s...", deployWorkflow, versionTag)
+	if err := dispatchWorkflow(deployRepo, deployWorkflow, map[string]string{"version_tag": versionTag}); err != nil {
 		log.Fatalf("Failed to dispatch deploy workflow: %v", err)
 	}
 
 	log.Info("Waiting for deploy workflow to start...")
-	deployRun, err := waitForNewRun(wikiDeployRepo, wikiDeployWorkflow, "workflow_dispatch", "", priorRunID)
+	deployRun, err := waitForNewRun(deployRepo, deployWorkflow, "workflow_dispatch", "", priorRunID)
 	if err != nil {
 		log.Fatalf("Failed to find dispatched deploy run: %v", err)
 	}
@@ -157,7 +171,7 @@ func runDeploy(versionTag string, noWait bool) {
 		return
 	}
 
-	if err := waitForRunCompletion(wikiDeployRepo, deployRun.DatabaseID, wikiDeployPollLimit, "deploy"); err != nil {
+	if err := waitForRunCompletion(deployRepo, deployRun.DatabaseID, wikiDeployPollLimit, "deploy"); err != nil {
 		log.Fatalf("Deploy did not complete successfully: %v", err)
 	}
 	log.Info("Deploy completed successfully.")
